Ratpack version 1.4 is now supported
