+++ instrumentation/ratpack-1.4/ratpack-1.4.gradle
@@ -10,7 +10,7 @@ muzzle {
   pass {
     group = "io.ratpack"
     module = 'ratpack-core'
-    versions = "[1.5.0,)"
+    versions = "[1.4.0,)"
     skipVersions += ["0.9.12", "0.9.13", "0.9.14",]
   }
 }
@@ -22,10 +22,10 @@ testSets {
 }
 
 dependencies {
-  compileOnly group: 'io.ratpack', name: 'ratpack-core', version: '1.5.0'
+  compileOnly group: 'io.ratpack', name: 'ratpack-core', version: '1.4.0'
 
   testImplementation project(':instrumentation:netty:netty-4.1')
-  testImplementation group: 'io.ratpack', name: 'ratpack-groovy-test', version: '1.5.0'
+  testImplementation group: 'io.ratpack', name: 'ratpack-groovy-test', version: '1.4.0'
   latestDepTestImplementation group: 'io.ratpack', name: 'ratpack-groovy-test', version: '+'
   if (JavaVersion.current().isCompatibleWith(JavaVersion.VERSION_11)) {
     testImplementation group: 'com.sun.activation', name: 'jakarta.activation', version: '1.2.2'
