+++ instrumentation/ratpack-1.4/src/main/java/io/opentelemetry/auto/instrumentation/ratpack/DefaultExecutionInstrumentation.java
@@ -16,68 +16,61 @@
 
 package io.opentelemetry.auto.instrumentation.ratpack;
 
-import static io.opentelemetry.auto.tooling.ClassLoaderMatcher.hasClassesNamed;
-import static io.opentelemetry.auto.tooling.bytebuddy.matcher.AgentElementMatchers.implementsInterface;
 import static java.util.Collections.singletonMap;
 import static net.bytebuddy.matcher.ElementMatchers.nameStartsWith;
 import static net.bytebuddy.matcher.ElementMatchers.named;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 
 import com.google.auto.service.AutoService;
-import com.google.common.net.HostAndPort;
 import io.opentelemetry.auto.tooling.Instrumenter;
 import java.util.Map;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
-import ratpack.func.Block;
+import ratpack.exec.internal.Continuation;
+import ratpack.func.Action;
 import ratpack.path.PathBinding;
 
 @AutoService(Instrumenter.class)
-public final class ContinuationInstrumentation extends Instrumenter.Default {
+public final class DefaultExecutionInstrumentation extends Instrumenter.Default {
 
-  public ContinuationInstrumentation() {
+  public DefaultExecutionInstrumentation() {
     super("ratpack");
   }
 
   @Override
-  public ElementMatcher<ClassLoader> classLoaderMatcher() {
-    // Optimization for expensive typeMatcher.
-    return hasClassesNamed("ratpack.exec.internal.Continuation");
-  }
-
-  @Override
   public ElementMatcher<? super TypeDescription> typeMatcher() {
-    return nameStartsWith("ratpack.exec.")
-        .<TypeDescription>and(implementsInterface(named("ratpack.exec.internal.Continuation")));
+    return named("ratpack.exec.internal.DefaultExecution");
   }
 
   @Override
   public String[] helperClassNames() {
-    return new String[] {packageName + ".BlockWrapper"};
+    return new String[] {packageName + ".ActionWrapper"};
   }
 
   @Override
   public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
     return singletonMap(
-        named("resume").and(takesArgument(0, named("ratpack.func.Block"))),
-        ContinuationInstrumentation.class.getName() + "$ResumeAdvice");
+        nameStartsWith("delimit") // include delimitStream
+            .and(takesArgument(0, named("ratpack.func.Action")))
+            .and(takesArgument(1, named("ratpack.func.Action"))),
+        DefaultExecutionInstrumentation.class.getName() + "$DelimitAdvice");
   }
 
-  public static class ResumeAdvice {
+  public static class DelimitAdvice {
 
     @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void wrap(@Advice.Argument(value = 0, readOnly = false) Block block) {
-      block = BlockWrapper.wrapIfNeeded(block);
+    public static void wrap(
+        @Advice.Argument(value = 0, readOnly = false) Action<? super Throwable> onError,
+        @Advice.Argument(value = 1, readOnly = false) Action<? super Continuation> segment) {
+      onError = ActionWrapper.wrapIfNeeded(onError);
+      segment = ActionWrapper.wrapIfNeeded(segment);
     }
 
-    public void muzzleCheck(final PathBinding binding, final HostAndPort host) {
+    public void muzzleCheck(final PathBinding binding) {
       // This was added in 1.4.  Added here to ensure consistency with other instrumentation.
       binding.getDescription();
-
-      // This is available in Guava 20 which was required starting in 1.5
-      host.getHost();
     }
   }
 }
