+++ instrumentation/ratpack-1.4/src/main/java/io/opentelemetry/auto/instrumentation/ratpack/ServerErrorHandlerInstrumentation.java
