+++ settings.gradle
@@ -129,7 +129,7 @@ include ':instrumentation:play-ws:play-ws-2.0'
 include ':instrumentation:play-ws:play-ws-2.1'
 include ':instrumentation:play-ws:play-ws-common'
 include ':instrumentation:rabbitmq-2.7'
-include ':instrumentation:ratpack-1.5'
+include ':instrumentation:ratpack-1.4'
 include ':instrumentation:reactor-3.1'
 include ':instrumentation:rediscala-1.8'
 include ':instrumentation:rmi'
